+++ packages/apiclient/src/fetchApiConnector.ts
@@ -40,9 +40,9 @@ export async function handleErrorStatusMiddleware(response: Response) {
  * custom values through the constructor.
  */
 export class FetchApiConnector implements IApiConnector {
-  private fetchFunction: FetchFunction;
-  private jsonSerializer: JsonSerializer;
-  private middleware: Middleware;
+  protected fetchFunction: FetchFunction;
+  protected jsonSerializer: JsonSerializer;
+  protected middleware: Middleware;
 
   constructor(configuration?: { fetchFunction?: FetchFunction; jsonSerializer?: JsonSerializer; middleware?: Middleware }) {
     this.fetchFunction = configuration?.fetchFunction ?? bind(window.fetch, window);
@@ -54,42 +54,31 @@ export class FetchApiConnector implements IApiConnector {
     return this.fetchFunction(url, { ...params, method: "GET" }).then(this.middleware);
   }
 
-  postText(url: string, text: string, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "POST", body: text }).then(this.middleware);
+  post(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
+    return this.fetchFunction(url, { ...params, method: "POST", body }).then(this.middleware);
+  }
+  postJson(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.post(url, this.jsonSerializer(content), appendJsonHeader(params));
   }
 
-  postJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.postText(url, this.jsonSerializer(content), appendJsonHeader(params));
-  }
-  postFormData(url: string, data: FormData, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "POST", body: data }).then(this.middleware);
+  put(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
+    return this.fetchFunction(url, { ...params, method: "PUT", body }).then(this.middleware);
   }
   putJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.putText(url, this.jsonSerializer(content), appendJsonHeader(params));
-  }
-  putText(url: string, text: string, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "PUT", body: text }).then(this.middleware);
-  }
-  putFormData(url: string, data: FormData, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "PUT", body: data }).then(this.middleware);
-  }
-  patchJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.patchText(url, this.jsonSerializer(content), appendJsonHeader(params));
-  }
-  patchText(url: string, text: string, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "PATCH", body: text }).then(this.middleware);
-  }
-  patchFormData(url: string, data: FormData, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "PATCH", body: data }).then(this.middleware);
+    return this.put(url, this.jsonSerializer(content), appendJsonHeader(params));
   }
 
-  delete(url: string, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "DELETE" }).then(this.middleware);
+  patch(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
+    return this.fetchFunction(url, { ...params, method: "PATCH", body }).then(this.middleware);
+  }
+  patchJson(url: string, content: any, params?: RequestInit): Promise<Response> {
+    return this.patch(url, this.jsonSerializer(content), appendJsonHeader(params));
+  }
+
+  delete(url: string, body?: BodyInit, params?: RequestInit): Promise<Response> {
+    return this.fetchFunction(url, { ...params, method: "DELETE", body }).then(this.middleware);
   }
   deleteJson(url: string, content: any, params?: RequestInit): Promise<Response> {
-    return this.deleteText(url, this.jsonSerializer(content), appendJsonHeader(params));
-  }
-  deleteText(url: string, text: string, params?: RequestInit): Promise<Response> {
-    return this.fetchFunction(url, { ...params, method: "DELETE", body: text }).then(this.middleware);
+    return this.delete(url, this.jsonSerializer(content), appendJsonHeader(params));
   }
 }