+++ packages/apiclient/src/types.ts
@@ -1,19 +1,15 @@
 export interface IApiConnector {
   get(url: string, params?: RequestInit): Promise<Response>;
 
+  post(url: string, body?: BodyInit, params?: RequestInit): Promise<Response>;
   postJson(url: string, content: any, params?: RequestInit): Promise<Response>;
-  postText(url: string, text: string, params?: RequestInit): Promise<Response>;
-  postFormData(url: string, data: FormData, params?: RequestInit): Promise<Response>;
 
+  put(url: string, body?: BodyInit, params?: RequestInit): Promise<Response>;
   putJson(url: string, content: any, params?: RequestInit): Promise<Response>;
-  putText(url: string, text: string, params?: RequestInit): Promise<Response>;
-  putFormData(url: string, data: FormData, params?: RequestInit): Promise<Response>;
 
+  patch(url: string, body?: BodyInit, params?: RequestInit): Promise<Response>;
   patchJson(url: string, content: any, params?: RequestInit): Promise<Response>;
-  patchText(url: string, text: string, params?: RequestInit): Promise<Response>;
-  patchFormData(url: string, data: FormData, params?: RequestInit): Promise<Response>;
 
-  delete(url: string, params?: RequestInit): Promise<Response>;
-  deleteText(url: string, text: string, params?: RequestInit): Promise<Response>;
+  delete(url: string, body?: BodyInit, params?: RequestInit): Promise<Response>;
   deleteJson(url: string, content: any, params?: RequestInit): Promise<Response>;
 }