+++ packages/datascreens/src/detailViewModel.ts
@@ -1,12 +1,12 @@
 import { BusyWatcher, ScreenBase } from "@frui.ts/screens";
 import { action, observable } from "mobx";
 
-export default abstract class DetailViewModel<TDetail> extends ScreenBase {
+export default abstract class DetailViewModel<TEntity> extends ScreenBase {
   busyWatcher = new BusyWatcher();
-  @observable item: TDetail;
+  @observable item: TEntity;
 
   @action.bound
-  setItem(item: TDetail) {
+  setItem(item: TEntity) {
     this.item = item;
   }
 
@@ -14,5 +14,5 @@ export default abstract class DetailViewModel<TDetail> extends ScreenBase {
     const item = await this.loadDetail();
     this.setItem(item);
   }
-  protected abstract loadDetail(): Promise<TDetail> | TDetail;
+  protected abstract loadDetail(): Promise<TEntity> | TEntity;
 }